+++ packages/react/src/index.ts
@@ -1,6 +1,7 @@
 "use client";
 
 export * from "./provider";
+export * from "./create-theme-root";
 export * from "./hooks";
 export * from "./theme-mode-button";
 export * from "./inspector";